Move stats from user model to a concern
